requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.129
botocore==1.29.129
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.89.1
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.6.0
jmespath==1.0.1
lxml==4.9.2
multidict==6.0.4
numpy==1.21.6
playwright==1.33.0
pydantic==1.10.4
pyee==9.0.4
python-dateutil==2.8.2
python-dotenv==0.21.1
python-http-client==3.3.7
PyYAML==6.0
redis==4.5.4
s3transfer==0.6.1
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starkbank-ecdsa==2.2.0
starlette==0.22.0
typing-extensions==4.4.0
urllib3==1.26.15
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.9.2
zipp==3.15.0
